Create exercise1.py
we get n and random words.
we create list and dictionary
firstly we add words to list
and then add words from list to dictionary, so we convert list to dictionary. then we take length of dictionary which gives us how many different words we have and we take values as how many times each word has been repeated.